# src/app/[locale]/api/partner/rentalcarmanager/openapi.yaml
openapi: 3.0.0
info:
  title: Rental Car Manager API
  version: 1.0.0
  description: API for integrating with the Rental Car Manager system.
servers:
  - url: /api/partner/rentalcarmanager
    description: Local proxy server
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: query
      name: accountKey # The name of the query parameter for the API key
...
paths:
...
  /cancelbooking:
    get:
      summary: Cancelling a booking (Agent API only).
      operationId: cancelBooking
      parameters:
        - in: query
          name: Rcm_reference_no
          schema:
            type: string
          required: true
          description: The unique Reference Key for the booking/quote.
        - in: query
          name: CancelReasonID
          schema:
            type: integer
          required: true
          description: ID from cancelreasons method.
      responses:
        '200':
          description: Successful response after cancelling booking.
          content:
            application/json:
              schema:
                type: object # Replace with actual schema if known
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /sendemail:
    get:
      summary: Send Confirmation Email.
      operationId: sendConfirmationEmail
      parameters:
        - in: query
          name: rcmReservationRef
          schema:
            type: string
          required: true
          description: The booking reference number.
      responses:
        '200':
          description: Successful response after sending email.
          content:
            application/json:
              schema:
                type: object # Replace with actual schema if known
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  responses:
    BadRequest:
      description: Bad request. Missing or invalid parameters.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: string
...
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string